requirements.txt
flask
gitpython
boto3
langchain
langchain-community
langchain-aws
faiss-cpu
